ai_service/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
confluent-kafka
# Optimizing for CPU to reduce image size and avoid IO errors during build
--extra-index-url https://download.pytorch.org/whl/cpu
torch
openai-whisper

boto3==1.34.0
requests==2.31.0

google-generativeai==0.8.3
openai==1.30.0
tiktoken==0.7.0